+++ serving-core/src/main/scala/com/stratio/sparta/serving/core/constants/AppConstant.scala
@@ -143,7 +143,7 @@ object AppConstant extends ZookeeperUtils {
   val DefaultGlobalParameters = Seq(
     new ParameterVariable("DEFAULT_OUTPUT_FIELD", "raw"),
     new ParameterVariable("DEFAULT_DELIMITER", ","),
-    new ParameterVariable("SPARK_EXECUTOR_BASE_IMAGE","qa.stratio.com/stratio/spark-stratio-driver:2.2.0-1.0.0"),
+    new ParameterVariable("SPARK_EXECUTOR_BASE_IMAGE","qa.stratio.com/stratio/spark-stratio-driver:2.2.0-2.0.0-ae1b428"),
     new ParameterVariable("SPARK_DRIVER_JAVA_OPTIONS","-Dconfig.file=/etc/sds/sparta/spark/reference.conf -XX:+UseConcMarkSweepGC -Dlog4j.configurationFile=file:///etc/sds/sparta/log4j2.xml"),
     new ParameterVariable("SPARK_EXECUTOR_EXTRA_JAVA_OPTIONS","-XX:+UnlockExperimentalVMOptions -XX:+UseCGroupMemoryLimitForHeap -XX:+UseConcMarkSweepGC"),
     new ParameterVariable("SPARK_STREAMING_CHECKPOINT_PATH","sparta/checkpoint"),